bootscript: make bootscript conditional on TARGET_BOOTLOADER_IMAGE
Commit 92ff85e8 made creation of the boot script conditional on
u-boot.imx. It did this in order to ensure that the tool mkimage
has been built, removing the need to install u-boot-tools on
the build machine.

The trouble is that the latest build structure can build multiple
U-Boot images for Quad Core, Solo, and various memory configurations.

Using TARGET_BOOTLOADER_IMAGE should achieve the same thing without
a reference to u-boot.imx (which is ambiguous at best).

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>