nitrogen6x: bootscript: allow setting of the WL127x mac address
This update allows the mac address for the WL12xx WiFi module to be set
in U-Boot in the 'wlmac' environment variable, which is translated into
the 'wl12xx.mac=' kernel command-line parameter.

Usage is as follows:
	U-Boot > setenv wlmac 00:19:b8:01:02:03
	U-Boot > saveenv && boot

Without something like this, you need to re-configure the mac
address in the wl1271-fw-2.bin file (offset 0xa370) using the
calibrator tool or a binary editor.

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>