Add mkimage.sh utility script
This script allows a disk image to be created from a working
build tree with a specified size.

Usage is:
	~/myandroid$ sudo device/boundary/mkimage.sh myandroid.img 3600 nitrogen6x

The first parameter is the output file name and the second is the
size (in MiBytes) of the resulting image.

The image will partitioned as with the mksdcard.sh script:
	p1	20 MiB boot
	p2	20 MiB recovery
	p5	512 MiB system
	p6	512 MiB cache
	p7	16 MiB vendor
	p8	16 MiB misc
	p4	variable size from end of the above to size specified

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>