nitrogen6x: fix BLE advertising for Silex module
The BLE local name needs to be reduced, otherwise the adv frame is
too large to be sent.

The bluedroid project has been modified to look for the new
BOARD_SUPPORTS_BLE_VND which allows to set the default BLE_VND_INCLUDED
value to true when set. This avoids having to change the
bdroid_builcfg.h header for each module.

Tested with Android sample app on Nitrogen6x:
https://github.com/googlesamples/android-BluetoothAdvertisements

The scanning device (smartphone) was using the other Android sample app:
https://github.com/googlesamples/android-BluetoothLeGatt

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>