nitrogen6x: add wlan mac address property for Silex
The Silex driver requires the MAC address to be passed as module
parameter. However this is not handled by Android WiFi HAL by default.

It works for wlcore since this part of TI driver is built-in.
But for modules it needs to be passed manually during the insmod
command.

The HAL has therefore been modified to take this into account.
It couldn't read the cmdline direclty since it is restricted to
root/radio user only, which is why we use a property instead.

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>