audio:  fix no audio output issue in tv platform [2/2]
PD#SWPL-773

Problem:
when SHIPPING API 28 is enabled,tv platform can not access
some system prop,and wrong configured to ott 16 2ch mode,
which cause no sound output.

Solution:
We use a Variable in tv_product.mk to define the TV product,
and in the audio hal makefile,we will read that variable to
determine if we need 32 bit 8 ch ouput or 2 ch 16 bit output.

Verify:
    compile  pass verified using Darwin and Franklin

Change-Id: Ia21a080920404701763fe3f525743fcd16f6e3ba
Signed-off-by: Jian Xu <jian.xu@amlogic.com>