nitrogen8m: bootscript: fix possible dtb alignment issue
As Google says [1]:
mkdtimg does not handle alignment for .dtb/.dtbo files but rather
appends them to the image.

So if a dtb is not 4-byte aligned, then the next one will have a wrong
start address. So add a simple copy of the dtb found from dtbo image to
a known-to-be-working/aligned address.

[1] https://source.android.com/devices/architecture/dto/partitions

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>